+++ changelog.md
@@ -19,6 +19,7 @@ Codion Change Log
 - EntityQueryModel.SelectAttributes added, returned by attributes().
 - EntityQueryModel.createSelect() added, DefaultEntityQueryModel refactored.
 - EntityQueryModel.query() renamed dataSource().
+- EntityQueryModel no longer implements Supplier, get() renamed query().
 ### is.codion.swing.framework.ui
 - EntityDialogs.selectionDialog() parameter now EntityTablePanel.
 - EntitySelectionDialogBuilder.dialogSize() and configureTablePanel() removed, includeSearchButton() added.